Release v0.9.0
This ends a long time of not really tagging versions (as it didn't
make a lot of sense - we could have tagged every commit with a new
minor release essentially). I still worked through the git log and GH
issues to create a nice CHANGELOG.md.

As dependency management tools really love version tags these days, I
plan to tag more often from now on. The ideas behind the pre-1.0
version numbers are explained in the README.md as changed in this
commit.

I also changed the issue template accordingly.

Signed-off-by: beorn7 <beorn@soundcloud.com>